functions/requirements.txt
azure-functions
azure-storage-queue>=12.0.0
twilio>=8.0.0